Address row comments
